requirements/prod.txt
alembic==1.6.4; python_version < '3.0'
alembic==1.7.5; python_version >= '3.0'
aniso8601==2.0.0
arrow==0.17.0
asn1crypto==0.24.0
attrs==20.2.0; python_version < '3.0'
attrs==21.2.0; python_version >= '3.0'
git+https://github.com/closeio/authalligator-client.git@fe93c9d2333d2949e44c48a2dd0a9a266734e026#egg=authalligator_client
backports.functools-lru-cache==1.4; python_version < '3.3'
backports.shutil_get_terminal_size==1.0; python_version < '3.3'
backports.ssl==0.0.9; python_version < '3.0'
backcall==0.2.0; python_version > '3.0'
jedi==0.18.1; python_version > '3.0'
boto==2.49.0
boto3==1.17.112; python_version < '3.0'
boto3==1.19.4; python_version >= '3.0'
botocore==1.20.112; python_version < '3.0'
botocore==1.22.9; python_version >= '3.0'
cchardet==2.1.1; python_version < '3.0'
cchardet==2.1.7; python_version >= '3.0'
certifi==2020.4.5.1; python_version < '3.0'
certifi==2021.10.8; python_version >= '3.0'
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.7; python_version >= '3.5'
ciso8601==2.2.0
click==7.1.2
colorlog==5.0.1; python_version < '3.0'
colorlog==6.5.0; python_version >= '3.0'
configparser==4.0.2; python_version < '3.0'
contextlib2==0.5.5; python_version < '3.0'
cryptography==3.3.2; python_version < '3.0'
cryptography==36.0.0; python_version >= '3.0'
decorator==4.4.2; python_version < '3.0'
decorator==5.1.0; python_version >= '3.0'
dnspython==1.16.0; python_version < '3.0'
dnspython==2.1.0; python_version >= '3.0'
enum34==1.1.3; python_version < '3.4'
faulthandler==3.2; python_version < '3.3'
filelock==3.4.0; python_version >= '3.0'
git+https://github.com/closeio/flanker-new.git@fa272d95345d45e8bb1f9bc32bc2c074bf52a484#egg=flanker
tld==0.12.6
Flask==1.1.4
Flask-RESTful==0.3.9
future==0.18.2
futures==2.2.0; python_version < '3.0'
gdata==2.0.18
gevent==21.8.0
gevent-openssl==1.2; python_version < '3.0'
greenlet==1.1.2
gunicorn==19.10.0; python_version < '3.0'
gunicorn==20.1.0; python_version >= '3.0'
guppy==0.1.10; python_version < '3.0'
guppy3==3.1.2; python_version >= '3.0'
hiredis==1.1.0
html2text==2019.8.11; python_version < '3.0'
html2text==2020.1.16; python_version >= '3.0'
icalendar==4.0.9
idna==2.10; python_version < '3.0'
idna==3.3; python_version >= '3.0'
imapclient==2.2.0
importlib-metadata==2.1.1; python_version < '3.0'
importlib-metadata==4.8.1; python_version >= '3.0'
importlib-resources==5.4.0; python_version >= '3.0'
importlib-resources==3.3.1; python_version < '3.0'
ipaddress==1.0.23; python_version < '3.3'
ipython==5.8.0; python_version < '3.0'
ipython==7.16.0; python_version >= '3.0'
ipython_genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.9.3
limitlion==0.10.0
lxml==4.6.4
Mako==1.0.7
MarkupSafe==1.1.1
mysqlclient==1.3.14
ndg-httpsclient==0.4.3; python_version < '3.0'
parso==0.8.2; python_version >= '3.0'
pathlib2==2.3.6; python_version < '3.0'
pexpect==4.8.0
pickleshare==0.7.5
ply==3.11
psutil==5.8.0
ptyprocess==0.7.0
prompt-toolkit==1.0.18; python_version < '3.0'
prompt-toolkit==3.0.22; python_version >= '3.0'
pyasn1==0.2.3; python_version < '3.0'
pycparser==2.21
pygments==2.5.2; python_version < '3.0'
pygments==2.10.0; python_version >= '3.0'
pyinstrument==3.2.0
pyinstrument-cext==0.2.2
pymongo==2.9.5  # For json_util in bson
Pympler==0.9
PyNaCl==1.4.0
pyOpenSSL==17.5.0; python_version < '3.0'
python-editor==1.0.4; python_version < '3.0'
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
redis==2.10.6
regex==2021.3.17; python_version < '3.0'
regex==2021.11.10; python_version >= '3.0'
requests==2.26.0
requests-file==1.5.1
rollbar==0.16.2
s3transfer==0.4.2; python_version < '3.0'
s3transfer==0.5.0; python_version >= '3.0'
scandir==1.10.0; python_version < '3.0'
setproctitle==1.1.10; python_version < '3.0'
setproctitle==1.2.2; python_version >= '3.0'
setuptools==44.0.0
simplegeneric==0.8.1; python_version < '3.0'
singledispatch==3.4.0.3; python_version < '3.0'
six==1.16.0
sqlalchemy==1.4.26
statsd==3.3.0
tldextract==2.2.3; python_version < '3.0'
tldextract==3.1.2; python_version >= '3.0'
structlog==20.1.0; python_version < '3.0'
structlog==21.4.0; python_version >= '3.0'
traitlets==4.3.3
typing==3.10.0.0; python_version < '3.5'
typing_extensions==4.0.1; python_version >= '3.0'
urllib3==1.26.7
vobject==0.9.6.1
wcwidth==0.2.5
WebOb==1.8.7
Werkzeug==1.0.1
zipp==1.2.0; python_version < '3.0'
zipp==3.6.0; python_version >= '3.0'
zope.event==4.5.0
zope.interface==5.4.0 